Api error handling

+++ api/router/users/index.js
@@ -5,39 +5,53 @@ const Heritage = require("../heritage/schema");
 const router = Router();
 
 router.get("/", async (req, res) => {
-  const { email, password } = req.query;
+  const { username, password } = req.query;
 
-  const isLogged = await User.findOne({ email });
-  if (!isLogged) {
-    return res.status(400).json({
-      errors: { emailError: "User isn't logged..." },
-    });
-  }
+  // Checking inputs
 
-  const user = await User.findOne({ email, password });
+  if(!username) return res.status(400).send({
+    errors: { username: "Username is required..." },
+  })
+
+  if(!password) return res.status(400).send({
+    errors: { password: "Password is required..." },
+  })
+
+  // Checking if the user isn't logged.
+  const user = await User.findOne({ username });
   if (!user) {
-    return res.status(400).json({
-      errors: { passwordError: "Password is incorrect..." },
+    return res.status(404).send({
+      errors: { username: "User isn't logged..." },
     });
   }
 
-  res.status(200).json({ userId: user._id });
-});
+  // Checking if the password is incorrect.
+  if(user.password != password) return res.status(404).send({
+    errors: { password: "Password is incorrect..." },
+  })
+
+  res.status(200).send({ userId: user._id });
+})
 
 router.get("/:myId", async (req, res) => {
   const { myId } = req.params;
+  if(myId === "undefined") return res.status(400).send({
+    errors: { id: "User's id is required..." }
+  })
   const myUser = await User.findById(myId)
     .populate("linkedAccounts")
     .populate("linkRequests")
     .populate("heritage");
 
-  const {
-    _doc: { password, ...data },
-  } = myUser;
+  if(!myUser) return res.status(400).send({
+    errors: { id: "Id didn't match any user..." }
+  });
 
-  // Changing each user's last connection date for how much time he/she has been offline.
+  const userData = myUser._doc;
+
+  // Changing each linked user's last connection date for how much time he/she has been offline.
   const currentDate = new Date().getTime();
-  data.linkedAccounts = data.linkedAccounts.map((linkedAccount) => {
+  userData.linkedAccounts = userData.linkedAccounts.map((linkedAccount) => {
     let months = parseInt(
       ((currentDate - linkedAccount.lastConnection) / 2629800000).toFixed(0)
     );
@@ -108,31 +122,58 @@ router.get("/:myId", async (req, res) => {
     return { ...linkedAccount._doc, lastConnection };
   });
 
-  res.status(200).send(data);
+  res.status(200).send(userData);
 });
 
 router.post("/", async (req, res) => {
   const { body } = req;
-  const data = { lastConnection: new Date(), linkedRequests: [] };
+  const userData = { lastConnection: new Date(), linkedRequests: [] };
 
-  if (body.email) data.email = body.email;
+  // Checking username possible errors.
+  if(!body.username) return res.status(400).send({
+    errors: { username: "Username is required..." },
+  });
+  if(body.username.length < 6) return res.status(400).send({
+    errors: { username: "Username is too short..." },
+  });
+  userData.username = body.username;
 
-  if (body.password) data.password = body.password;
+  // 400 status code in case the user already exists.
+  const userAlreadyExists = await User.findOne({ username: body.username });
+  if(userAlreadyExists) return res.status(400).send({
+    errors: { username: "User already exists..." },
+  });
 
-  if (body.heritage) {
-    const heritage = await Heritage({ amount: body.heritage });
-    heritage.save();
-    data.heritage = heritage._id;
-  }
+  // Checking password possible errors.
+  if(!body.password) return res.status(400).send({
+    errors: { password: "Password is required..." },
+  });
+  if(body.password.length < 6) return res.status(400).send({
+    errors: { password: "Password is too short..." },
+  });
+  userData.password = body.password;
+  
+  // Checking heritage possible errors. (and creating it if there aren't any)
+  if(!body.heritage) return res.status(400).send({
+    errors: { heritage: "Heritage is required..." },
+  });
+  const heritage = await Heritage({ amount: body.heritage });
+  heritage.save();
+  userData.heritage = heritage._id;
 
+  // idToBeLinked to link another user to my account. (if it exists)
   if (body.idToBeLinked) {
     const userToBeLinked = await User.findById(body.idToBeLinked);
-    if (userToBeLinked) data.linkedAccounts = [body.idToBeLinked];
+    if(!userToBeLinked) return res.status(404).send({
+      errors: { id: "Id to be linked didn't match any user..." }
+    })
+    if (userToBeLinked) userData.linkedAccounts = [body.idToBeLinked];
   }
 
-  const newUser = await User(data);
+  const newUser = await User(userData);
   newUser.save();
 
+  // idToBeLinked to link my own account to another one. (if it exists)
   if (body.idToBeLinked) {
     const userToBeLinked = await User.findById(body.idToBeLinked);
     await User.findByIdAndUpdate(userToBeLinked._id, {
@@ -142,36 +183,82 @@ router.post("/", async (req, res) => {
 
   res
     .status(201)
-    .json({ userId: body.idToBeLinked ? body.idToBeLinked : newUser._id });
+    .send({ userId: body.idToBeLinked ? body.idToBeLinked : newUser._id });
 });
 
 router.put("/:myId", async (req, res) => {
-  const {
-    body,
-    params: { myId },
-  } = req;
-  const data = {};
+  const { body, params: { myId } } = req;
+  const userData = {};
 
-  if (body.email) data.email = body.email;
-  if (body.password) data.password = body.password;
-  if (body.lastConnection) data.lastConnection = body.lastConnection;
+  if(myId === "undefined") return res.status(400).send({
+    errors: { id: "User's id is required..." }
+  })
 
-  if (body.emailToBeLinked) {
-    const userToBeLinked = await User.findOne({ email: body.emailToBeLinked });
-    if (!userToBeLinked) {
-      return res.status(400).json({
-        errors: { emailError: "User doesn't exist..." },
+  const myUser = await User.findById(myId);
+
+  if(!myUser) return res.status(400).send({
+    errors: { id: "Id didn't match any user..." },
+  });
+
+  if (body.username) {
+    const userAlreadyExists = await User.findOne({ username: body.username });
+    if(userAlreadyExists) return res.status(400).send({
+      errors: { username: "User already exists..." },
+    });
+    if(body.username.length < 6) return res.status(400).send({
+      errors: { username: "Username is too short..." },
+    });
+    // userData.username assignment in case there weren't any exceptions.
+    userData.username = body.username;
+  }
+  if (body.password) {
+    if(body.password.length < 6) return res.status(400).send({
+      errors: { password: "Password is too short..." },
+    });
+    // userData.password assignment in case there weren't any exceptions.
+    userData.password = body.password;
+  }
+  if (body.lastConnection) userData.lastConnection = body.lastConnection;
+
+  if (body.usernameToBeLinked) {
+    const userToBeLinked = await User.findOne({ username: body.usernameToBeLinked });
+
+    if (!userToBeLinked) return res.status(400).send({
+      errors: { username: "User doesn't exist..." },
+    });
+
+    if(userToBeLinked._id == myId) return res.status(400).send({
+      errors: { username: "You can't link your own account..." },
+    });
+
+    for(linkedAccount of userToBeLinked.linkedAccounts) {
+      if(linkedAccount == myId) return res.status(400).send({
+        errors: { username: "You are already linked to this user..." },
       });
     }
+
+    for(linkRequest of userToBeLinked.linkRequests) {
+      if(linkRequest == myId) return res.status(400).send({
+        errors: { username: "You have already tried to link to this user..." },
+      });
+    }
+
+    // The link request is sent to the other user 
+    // in case there weren't any exceptions
+
     await User.findByIdAndUpdate(userToBeLinked._id, {
       linkRequests: [...userToBeLinked.linkRequests, myId],
     });
   }
 
+  // Unlink users logic.
   if (body.idToBeUnlinked) {
-    const myUser = await User.findById(myId);
     const anotherUser = await User.findById(body.idToBeUnlinked);
-    data.linkedAccounts = myUser.linkedAccounts.filter(
+    if(!anotherUser) return res.status(400).send({
+      errors: { id: "User to unlink wasn`t found..." },
+    });
+
+    userData.linkedAccounts = myUser.linkedAccounts.filter(
       (linkedAccount) => linkedAccount._id != body.idToBeUnlinked
     );
     const anotherUserLinkedAccounts = anotherUser.linkedAccounts.filter(
@@ -182,16 +269,19 @@ router.put("/:myId", async (req, res) => {
     });
   }
 
+  // Link users logic. (through 'link requests')
   if (body.linkUserResponse) {
-    const myUser = await User.findById(myId);
     const anotherUser = await User.findById(body.linkUserResponse.id);
-
-    data.linkRequests = myUser.linkRequests.filter((linkRequest) => {
+    if(!anotherUser) return res.status(400).send({
+      errors: { id: "User to link wasn`t found..." },
+    });
+    
+    userData.linkRequests = myUser.linkRequests.filter((linkRequest) => {
       return linkRequest != body.linkUserResponse.id;
     });
 
     if (body.linkUserResponse.accepted) {
-      data.linkedAccounts = [
+      userData.linkedAccounts = [
         ...myUser.linkedAccounts,
         body.linkUserResponse.id,
       ];
@@ -201,15 +291,22 @@ router.put("/:myId", async (req, res) => {
     }
   }
 
-  await User.findByIdAndUpdate(myId, data);
+  await User.findByIdAndUpdate(myId, userData);
 
-  res.status(200).send();
+  res.status(204).send();
 });
 
 router.delete("/:myId", async (req, res) => {
   const { myId } = req.params;
+  if(myId === "undefined") return res.status(400).send({
+    errors: { id: "User's id is required..." }
+  })
   const myDeletedUser = await User.findByIdAndDelete(myId);
+  if(!myDeletedUser) return res.status(400).send({
+    errors: { id: "Id didn't match any user..." }
+  })
 
+  // Remove my account from the other users' list of linked accounts.
   myDeletedUser.linkedAccounts.forEach(async (linkedAccount) => {
     const anotherUser = await User.findById(linkedAccount._id);
     const newLinkedAccounts = anotherUser.linkedAccounts.filter(
